# erasure-code/isa/xor_op: add neon-based region_xor implementation