--- a/src/erasure-code/isa/xor_op.h
+++ b/src/erasure-code/isa/xor_op.h
@@ -83,6 +83,7 @@ region_sse2_xor(char** src /* array of 64-byte aligned source pointer to xor */,
                 int src_size /* size of the source pointer array */,
                 unsigned size /* size of the region to xor */);
 
+#if defined(__aarch64__) && defined(__ARM_NEON)
 // -------------------------------------------------------------------------
 // compute region XOR like parity = src[0] ^ src[1] ... ^ src[src_size-1]
 // using NEON 32-byte operations
@@ -92,5 +93,5 @@ region_neon_xor(char** src    /* array of 64-byte aligned source pointer to xor 
                 char* parity  /* 32-byte aligned output pointer containing the parity */,
                 int src_size  /* size of the source pointer array */,
                 unsigned size /* size of the region to xor */);
-
+#endif // __aarch64__ && __ARM_NEON
 #endif // EC_ISA_XOR_OP_H