@Svelar 2023-12-15
> Same question as in #54884, does __ARM_NEON guarantee the presence of 2 load channels? I'm afraid I'm not super familiar with different ARM architectures and potential issues here. Thanks much!

Same answer as in #54884, __ARM_NEON doesn't control the number of LD channels, and the number would not affect functionality but only different performance.