@Svelar 2023-12-18
> Are there tests for this which would validate that we get the same answer with and without this optimization?

Cuz tests are included in src/test/erasure-code/TestErasureCode.cc, I didn't add more tests. IMO, both X86 and ARM make check passed guarantee this.