@Svelar 2024-01-08
https://pulpito.ceph.com/svelar-2024-01-05_01:31:59-rados:thrash-erasure-code-isa-wip-svelar-testing-1-2024-01-04-0653-distro-default-smithi/

1 Dead caused by infra issues. Seems like reasonable to merge from the runs.